% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREEBOOST.R
\name{TFBoost.predict}
\alias{TFBoost.predict}
\title{TFBoost.predict}
\usage{
TFBoost.predict(model, newx, newy, newz)
}
\arguments{
\item{model}{an object returned by TFBoost}

\item{newx}{functional predictor matrix for test data (matrix/dataframe)}

\item{newy}{scalar predictor matrix for test data (matrix/dataframe, optional, requires if model is trained with z_train and z_val)}

\item{newz}{scalar response vector for test data (vector/dataframe)}
}
\value{
A list with with the following components:

\item{f_t_test}{predicted values with model at the early stopping iteration using x_test (or x_test and z_test) as the predictors}
\item{err_test}{a vector of test errors before and at the early stopping iteration (returned if make_prediction = TRUE in control)}
\item{save_f_test}{a matrix of test function estimates at all iterations (returned if save_f = TRUE in control)}
}
\description{
A function to make predictions and calculate test error given an object returned by TFBoost and test data
}
\details{
A function to make predictions and calculate test error given an object returned by TFBoost and test data
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
