% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREEBOOST.R
\name{TFBoost}
\alias{TFBoost}
\title{Tree-based functional boosting}
\usage{
TFBoost(
  x_train,
  z_train = NULL,
  y_train,
  x_val,
  z_val = NULL,
  y_val,
  x_test,
  z_test = NULL,
  y_test,
  grid,
  t_range,
  niter = 10,
  control = TFBoost.control()
)
}
\arguments{
\item{x_train}{functional predictor matrix in the training data (matrix/dataframe)}

\item{z_train}{scalar predictor matrix in the training data (matrix/dataframe, optional)}

\item{y_train}{scalar response vector in the training data (vector/dataframe)}

\item{x_val}{functional predictor matrix in the validation data (matrix/dataframe)}

\item{z_val}{scalar predictor matrix in the validation data (matrix/dataframe, optonal)}

\item{y_val}{scalar response vectorin the validation data (vector/dataframe)}

\item{x_test}{functional predictor matrix for test data (matrix/dataframe, optional, required when \code{make_prediction} in control is \code{TRUE})}

\item{z_test}{scalar predictor matrix for test data (matrix/dataframe, optional, required when \code{make_prediction} in control is \code{TRUE})}

\item{y_test}{scalar response vector for test data (vector/dataframe, optional, required when \code{make_prediction} in control is \code{TRUE} and \code{z_train} and \code{z_val} are provided)}

\item{grid}{common grid that the \code{x_train}, \code{x_val}, and \code{x_test} are evaluated at (vector)}

\item{t_range}{domain of the functional predictor, provided in the form of c(left end of the domain, right end of the domain) (vector)}

\item{niter}{number of boosting iterations (numeric)}

\item{control}{a named list of control parameters, as returned by \code{\link{TFBoost.control}}}
}
\value{
A list with the following components:

\item{B}{predicted values with model at the early stopping iteration using x_test (or x_test and z_test) as the predictors}
\item{loss_train}{a vector of training errors for all iterations}
\item{loss_val}{a vector of validation errors for all iterations}
\item{f_train_t}{predicted values with model at the early stopping iteration on the training predictors}
\item{f_val_t}{predicted values with model at the early stopping iteration on the validation predictors}
\item{f_t_test}{predicted values with model at the early stopping iteration on the test predictors (returned if make_prediction = TRUE in control)}
\item{early_stop}{early stopping iteration}
\item{err_train}{a vector of training mean-squared-errors for all iterations}
\item{err_val}{a vector of validation mean-squared-errors for all iterations}
\item{err_test}{a vector of test mean-squared-errors before and at the early stopping iteration (returned if make_prediction = TRUE in control)}
\item{grid}{\code{grid} form the input arguments}
\item{t_range}{\code{t_range} from the input arguments}
\item{init_vals}{a constant to initialize the function estimates for training, validation, and test sets}
\item{tree_obj}{a list of fitted functional multi-index trees (one per iteration and returned from \link{TREE})}
\item{alpha}{a vector of boosting step sizes (one per iteration)}
\item{control}{\code{control} from the input arguments}
\item{save_f_train}{a matrix of training function estimates at all iterations (returned if save_f = TRUE in control)}
\item{save_f_val}{a matrix of validation function estimates at all iterations (returned if save_f = TRUE in control)}
\item{save_f_test}{a matrix of test function estimates before and at the early stopping iteration (returned if save_f = TRUE and make_prediction = TRUE in control)}
}
\description{
This function implements a tree-based boosting algorithm for functional regression,
}
\details{
This function implements  a tree-based boosting algorithm for functional regression (TFBoost).
This function uses the functions available in the \code{rpart} package to construct functional multi-index regression trees.
}
\author{
Xiaomeng Ju, \email{xiaomeng.ju@stat.ubc.ca}
}
