% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREE.R
\name{TREE.control}
\alias{TREE.control}
\title{Tuning and control parameters for the functional multi-index tree}
\usage{
TREE.control(
  make_prediction = TRUE,
  d = 1,
  minbucket = 2,
  tree_type = "a",
  nmulti = 3,
  nscreen = 50,
  num_dir = 20,
  num_index = 1
)
}
\arguments{
\item{make_prediction}{logical indicating whether to make predictions using \code{newx} (defaults to \code{TRUE})}

\item{d}{the maximum depth of the functional multi-index tree (numeric, defaults to 1)}

\item{minbucket}{the minimum number of observations per node of the functional multi-index tree (numeric, defaults to 2)}

\item{tree_type}{type of the functional multi-index tree, 'A' or 'B' (character, defaults to 'A')}

\item{nmulti}{number of random initial points  for performing optimization to fit a type A tree (numeric, defaults to 5)}

\item{nscreen}{number of random points from which to select the initial points to fit a type A tree (numeric, defaults to 30)}

\item{num_dir}{number of random directions sampled at every iteration for type b tree (numeric, defaults to 200)}

\item{num_index}{number of indices for type a tree (numeric, defaults to 1)}
}
\value{
A list of all input parameters
}
\description{
Tuning and control parameters for the functional multi-index tree.
}
\details{
Various tuning and control parameters for the functional multi-index tree algorithm implemented in the
function \code{\link{TREE}}
}
\author{
Xiaomeng Ju, \email{xiaomeng.ju@stat.ubc.ca}
}
