% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREE.R
\name{TREE.predict}
\alias{TREE.predict}
\title{TREE.predict}
\usage{
TREE.predict(model, newx, newz)
}
\arguments{
\item{model}{an object returned by TREE}

\item{newx}{matrix of the functional predictor's basis projections in the test data (matrix/dataframe)}

\item{newz}{matrix of the scalar predictors in the test data (matrix/dataframe, required if the model includes scalar predictors)}
}
\value{
\item{pred}{predicted values with model using x_test (or x_test and z_test) as the predictors}
}
\description{
A function to make predictions given an object returned by TREE and test data
}
\details{
A function to make predictions given an object returned by TREE and test data
}
\author{
Xiaomeng Ju, \email{xmengju@stat.ubc.ca}
}
