% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREE.R
\name{TREE}
\alias{TREE}
\title{Functional multi-index tree}
\usage{
TREE(x, y, z, newx, newy, newz, random.seed, control = TREE.control())
}
\arguments{
\item{x}{matrix of the functional predictor's basis projections in the training data (matrix/dataframe)}

\item{y}{response vector in the training data (vector/dataframe)}

\item{z}{matrix of the scalar predictors in the training data (matrix/dataframe)}

\item{newx}{matrix of the functional predictor's basis projections in the test data (matrix/dataframe, optional)}

\item{newy}{response vector in the test data (matrix/dataframe, optional)}

\item{newz}{matrix of the scalar predictors in the test data (matrix/dataframe, optional)}

\item{control}{a named list of control parameters, as returned by \code{\link{TREE.control}}}
}
\value{
A list with the following components:

\item{beta_opt}{estimated index coefficients for type A tree, returned if \code{tree_type = "A"}}
\item{betas_selected}{the selected random directions for type B tree, returned if \code{tree_type = "B"}}
\item{pred_train}{predicted values with model on the training predictors}
\item{pred_test}{predicted values with model on the test predictors, returned if \code{make_predictions = TRUE}}
\item{control}{\code{control} from the input arguments}
\item{tree.model}{fitted tree estimator, an rpart object}
}
\description{
This function implements a algorithm for functional multi-index tree.
}
\details{
This function implements a functional multi-index tree algorithm devloped based on functions available in the \code{rpart} package.
}
\author{
Xiaomeng Ju, \email{xiaomeng.ju@stat.ubc.ca}
}
